Pass objects and vars explicitly to the timer callback.
Previously the script used a very creative way to pass more arguments than FrameTimerWithArg can handle. It created multiple temp objects and used Z and direction fields as data storage for additional vars. And to pass multiple objects to the same callback and made an assumption that object IDs will be sequential if created in the same script call.

Now, the script passes created objects explicitly, as well as other required function dependencies.

Some data is still passed via objects, to make sure the script works the same way in case the temp object is deleted or persisted.